2010-11-11  Eric Carlson  <eric.carlson@apple.com>
        Build fix after r71824.

        * html/HTMLMediaElement.cpp:
        (WebCore::HTMLMediaElement::invalidateCachedTime): Use an explicit cast from double to float.
        (WebCore::HTMLMediaElement::currentTime): Ditto.


Canonical link: https://commits.webkit.org/62302@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71826 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/html/HTMLMediaElement.cpp
@@ -1226,7 +1226,7 @@ void HTMLMediaElement::invalidateCachedTime()
     // too early.
     static const float minimumTimePlayingBeforeCacheSnapshot = 0.5;
 
-    m_minimumWallClockTimeToCacheMediaTime = WTF::currentTime() + minimumTimePlayingBeforeCacheSnapshot;
+    m_minimumWallClockTimeToCacheMediaTime = static_cast<float>(WTF::currentTime()) + minimumTimePlayingBeforeCacheSnapshot;
     m_cachedTime = invalidMediaTime;
 }
 
@@ -1263,7 +1263,7 @@ float HTMLMediaElement::currentTime() const
 
         // Not too soon, use the cached time only if it hasn't expired.
         if (wallClockDelta < maximumDurationToCacheMediaTime) {
-            float adjustedCacheTime = m_cachedTime + (m_playbackRate * wallClockDelta);
+            float adjustedCacheTime = static_cast<float>(m_cachedTime + (m_playbackRate * wallClockDelta));
 
 #if LOG_CACHED_TIME_WARNINGS
             float delta = adjustedCacheTime - m_player->currentTime();
